Agent concurrency (#21)
When using two different instances of the API at the same time, an error occurred: Error calling agent. Status: 500. Details: {"detail":"<asyncio.locks.Event object at 0x3e3dd1bc6250 [unset]> is bound to a different event loop"}

This was because the agent instance was generated only once, when the API was instanciated, now the agent instance is created everytime the ask_agent endpoint is called to avoid this issue.

Commits:

* Generate a function to get a new agent instance

* Modify the ask_agent endpoint to generate an agent instance to avoid loop event  errors