Solution to Update BigQuery Data Problem
Due to BigQuery was made especially for data analysis, operations such as update
the data cannot be done so easily on new data due to the original values of the
data are stored in a streaming buffer, which can take up to 90 minutes to
be truly inserted in the BigQuery table; until then, the data should be available
to be modified.

Nevertheless, as the chatbot requires to store data in streaming, the full chat history
could not be stored in the chat session table, as it cannot be quicly rewritten, but can
be quickly queried. To use BigQuery and take advantage of its free tier and the fast
retrieval feature, instead of moving to another database such as PostgreSQL, it was
decided to create two new tables which will help for data analysis and to restore the
chat history easily.